+++ src/text/str/namespace/mod.rs
@@ -3,15 +3,21 @@
 //! Defines the [`Str`] namespace.
 //
 
-mod definition; // Str
+#[cfg(test)]
+mod _tests;
+
+mod define; // Str
 
 // impls
+mod utf8_traversal; // UTF-8 conversion and traversal
+mod boundary; // equality and boundary
+mod writing; // writing and repetition
 mod range;
 mod take;
 mod split;
 
 crate::structural_mods! { // _mods
     _mods {
-        pub use super::definition::*;
+        pub use super::define::Str;
     }
 }