+++ src/text/translit/namespace.rs
@@ -8,6 +8,9 @@
 ///
 /// Provides small, mostly `const` mappings
 /// from Unicode scalars to simpler textual approximations.
+///
+/// # Features
+/// All current functionality depends on the `translit` feature.
 #[doc = crate::_doc_meta!{location("text")}]
 #[derive(Debug)]
 pub struct Translit;